requirements.txt
streamlit
google-genai
python-dotenv
reportlab
PyPDF2
pdfplumber
